requirements.txt
# ===== Core Dependencies =====
langgraph>=1.1.0
langgraph-supervisor>=0.0.30
langgraph-checkpoint>=2.0.0
langchain-deepseek>=0.1.0
langchain-core>=0.3.0
langchain-openai>=0.3.0

# ===== Data Analysis =====
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# ===== Web UI =====
streamlit>=1.30.0

# ===== Utilities =====
python-dotenv>=1.0.0
pydantic>=2.0.0

# ===== Testing =====
pytest>=7.0.0
pytest-asyncio>=0.21.0
